Modify the event test case
When this test case initially implemented,
the div element which isn't overflow but specified with 'overflow-x: scroll' was considered as "scroll container". Because of that the element was also "spatial navigation container".

But the definition of "scroll container" changed to "element which has overflowing content and specified with overflow CSS property". (related PR: https://github.com/WICG/spatial-navigation/pull/198)

Therefore, the element (#c1) needs to specified with `--spatial-navigation-contain: contain` for making it as the spatial navgation container for satisfying the initial intention of the test.